+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Access control menu available for MoSys grids
+
+### Changed
+
+- Moved ACL functionality to [mbjs-quasar](https://gitlab.rlp.net/motionbank/mbjs/quasar) module
+
+
+## [1.2.1] - 2019-02-15
+
 ### Fixed
 
 - Post-annotator timestamps now display correct values, independent of timezone or DST
 - currentIndex property no longer throws when attempting to scroll to invalid annotation index
 - Timestamps in post-annotator are no longer being cut off, display hours as well
+
 
 ## [1.2.0] - 2019-02-14
 
@@ -45,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - [mbjs-api-client](https://gitlab.rlp.net/motionbank/mbjs/api-client) now at version 1.1.0
+
 
 ## [1.1.1] - 2019-01-28
 
@@ -99,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Start proper versioning at 1.0.0
 
 
-[Unreleased]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.2.0...master
+[Unreleased]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.2.1...master
+[1.2.1]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.2.0...v1.2.1
 [1.2.0]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.1.1...v1.2.0
 [1.1.1]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.1.0...v1.1.1
 [1.1.0]: https://gitlab.rlp.net/motionbank/systems-frontend/compare/v1.0.0...v1.1.0